Fix some codes
