+++ src/main/java/com/study/dbtest/domain/enroll/dto/response/CountStudentWithSameNamePerCourseResDto.java
@@ -1,14 +1,12 @@
 package com.study.dbtest.domain.enroll.dto.response;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
 public class CountStudentWithSameNamePerCourseResDto {
     private Integer courseId;
     private String courseName;
