+++ src/main/java/com/study/dbtest/domain/enroll/dto/response/StudentCoursesResDto.java
@@ -1,7 +1,6 @@
 package com.study.dbtest.domain.enroll.dto.response;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -10,7 +9,6 @@ import java.util.Date;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
 public class StudentCoursesResDto {
     private Integer studentId;
     private String name;
