+++ src/main/java/com/study/dbtest/model/entity/Course.java
@@ -10,7 +10,6 @@ import java.util.List;
 @NoArgsConstructor
 @Builder
 @Entity
-@ToString
 public class Course {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
@@ -24,6 +23,6 @@ public class Course {
     @Column
     private Integer score;
 
-//    @OneToMany(mappedBy = "course", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+//    @OneToMany(mappedBy = "course", fetch = FetchType.LAZY)
 //    private List<Enrollment> enrollments;
 }
