+++ src/main/java/com/study/dbtest/model/repository/StudentRepository.java
@@ -12,6 +12,11 @@ import java.util.List;
 
 @Repository
 public interface StudentRepository extends JpaRepository<Student, Integer> {
-
-
+    @Query(value = "select new com.study.dbtest.domain.enroll.dto.response.StudentCoursesResDto(s.id, s.name, c.id, c.courseName, c.instructor, e.enrollmentDate) " +
+            "from Enrollment e " +
+            "join Student s on s.id = e.student.id " +
+            "join Course c on e.course.id = c.id " +
+            "where s.id = :studentId"
+    )
+    List<StudentCoursesResDto> findCoursesById(@Param("studentId") int id);
 }
